+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- Add `from_str()` function.
+- Add `spec::SecurityScheme::MutualTls` enum variant.
+- Add `spec::SecurityScheme::{ApiKey, Http, OAuth2, OpenIdConnect}::description` fields.
+- The `Spec::paths` field is now optional to closer align with the spec.
+- The `spec::Operation::responses` field is now optional to closer align with the spec.
+- The `spec::Schema::{exclusive_maximum, exclusive_minimum}` fields are now of type `Option<serde_json:Number>` to closer align with the spec.
+- Migrate YAML parsing to `serde_yml`. Exposed error type(s) have been altered.
+
 ## 0.5.0
 
 - Update `http` dependency to `1`.